+++ pkg/parsers/MAINTAINERS
@@ -1,0 +1,1 @@
+Erik Hollensbe <github@hollensbe.org> (@erikh)
